Reorg demo code, update content
