+++ src/demos/graph/2-graph-layout/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset='UTF-8' />
     <link rel="stylesheet" href="https://unpkg.com/react-vis@1.5.0/dist/style.css">
-    <title>Basic Graph Vis Tutorial - Graph Layout</title>
+    <title>Basic Graph Vis Tutorial - Interaction</title>
     <style>
       body {
         font-family: ff-clan-web-pro, "Helvetica Neue", Helvetica, sans-serif;
