+++ src/demos/graph/2-graph-layout/package.json
@@ -5,6 +5,7 @@
     "build": "NODE_ENV=production webpack --env.prod=true"
   },
   "dependencies": {
+    "d3-force": "^1.0.6",
     "deck.gl": "^4.1.0",
     "react": "^15.4.1",
     "react-dom": "^15.4.1"
