+++ src/demos/graph/2-graph-layout/src/app.js
@@ -1,84 +1,63 @@
 /* global window */
-import React, {Component} from 'react';
+import React, { Component } from 'react'
 
 // data
 import sampleGraph from '../../data/sample-graph';
 
 // utils
 import Graph from '../../common/graph';
+import LayoutEngine from '../../common/layout-engine';
 
 // components
-import GraphRender from './graph-render'
-
-function randomPosition(width, height) {
-  const halfWidth = width / 2;
-  const halfHeight = height / 2;
-  const x = Math.random() * width - halfWidth;
-  const y = Math.random() * height - halfHeight;
-  return [x, y];
-}
+import GraphRender from './graph-render';
 
 export default class App extends Component {
-
   constructor(props) {
     super(props);
     this.state = {
       viewport: {
-        width: window.innerWidth,
-        height: window.innerHeight
+        width: 1000,
+        height: 700
       },
       graph: new Graph()
     };
+    this._engine = new LayoutEngine();
   }
 
   componentDidMount() {
-    window.addEventListener('resize', this.onResize);
+    this._engine.registerCallbacks(this.reRender);
     this.processData();
   }
 
   componentWillUnmount() {
-    window.removeEventListener('resize', this.onResize);
+    this._engine.unregisterCallbacks();
   }
 
-  onResize = () => {
-    this.setState({
-      viewport: {
-        width: window.innerWidth,
-        height: window.innerHeight
-      }
-    });
-  }
-
-  processData() {
+  processData = () => {
     if (sampleGraph) {
-      const width = this.state.viewport.width;
-      const height = this.state.viewport.height;
       const newGraph = new Graph();
       sampleGraph.nodes.forEach(node =>
-        newGraph.addNode({
-          id: node.id,
-          position: randomPosition(width, height)
-        })
+        newGraph.addNode(node)
       );
       sampleGraph.edges.forEach(edge =>
         newGraph.addEdge(edge)
       );
       this.setState({graph: newGraph});
+      // update engine
+      this._engine.update(newGraph);
+      this._engine.start();
     }
   }
+
+  reRender = () => this.forceUpdate()
 
   // node accessors
   getNodeColor = node => [94, 94, 94]
   getNodeSize = node => 10
-  getNodePosition = node => node.position
 
   // edge accessors
   getEdgeColor = edge => [64, 64, 64]
   getEdgeWidth = () => 2
-  getEdgePosition = edge => ({
-    sourcePosition: this.state.graph.findNode(edge.source).position,
-    targetPosition: this.state.graph.findNode(edge.target).position
-  })
 
   render() {
     if (this.state.graph.isEmpty()) {
@@ -90,16 +69,18 @@ export default class App extends Component {
         /* viewport related */
         width={this.state.viewport.width}
         height={this.state.viewport.height}
+        /* update triggers */
+        positionUpdateTrigger={this._engine.alpha()}
         /* nodes related */
         nodes={this.state.graph.nodes}
         getNodeColor={this.getNodeColor}
         getNodeSize={this.getNodeSize}
-        getNodePosition={this.getNodePosition}
+        getNodePosition={this._engine.getNodePosition}
         /* edges related */
         edges={this.state.graph.edges}
         getEdgeColor={this.getEdgeColor}
         getEdgeWidth={this.getEdgeWidth}
-        getEdgePosition={this.getEdgePosition}
+        getEdgePosition={this._engine.getEdgePosition}
       />
     );
   }
