--- src/demos/graph/3-interactive-graph/index.html
@@ -1,30 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset='UTF-8' />
-    <link rel="stylesheet" href="https://unpkg.com/react-vis@1.5.0/dist/style.css">
-    <title>Basic Graph Vis Tutorial - Interaction</title>
-    <style>
-      body {
-        font-family: ff-clan-web-pro, "Helvetica Neue", Helvetica, sans-serif;
-        font-size: 12px;
-        line-height: 1.833;
-        margin: 0;
-        overflow: hidden;
-        padding: 0;
-      }
-      .intro {
-        align-items: center;
-        display: flex;
-        height: 100vh;
-        justify-content: center;
-        position: absolute;
-        width: 100vw;
-      }
-    </style>
-  </head>
-  <body>
-    <div id='root' />
-    <script type="text/javascript" src="bundle.js"></script>
-  </body>
-</html>
