--- src/demos/graph/4-final-version/README.md
@@ -1,9 +1,0 @@
-This is a minimal standalone version of the ArcLayer example
-on [deck.gl](http://deck.gl) website.
-
-### Usage
-Copy the content of this folder to your project. Run
-```
-npm install
-npm start
-```
