--- src/demos/graph/4-final-version/package.json
@@ -1,22 +1,0 @@
-{
-  "scripts": {
-    "start-local": "webpack-dev-server --env.local --progress --hot --port 3030 --open",
-    "start": "webpack-dev-server --progress --hot --port 3030 --open",
-    "build": "NODE_ENV=production webpack --env.prod=true"
-  },
-  "dependencies": {
-    "d3-force": "^1.0.6",
-    "deck.gl": "^4.1.0",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1"
-  },
-  "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "webpack": "^2.2.0",
-    "webpack-dev-server": "^2.2.0"
-  }
-}
