+++ src/demos/index.js
@@ -10,8 +10,8 @@ import GeospatialAppLinkingItAll from './building-a-geospatial-app/6-linking-it-
 
 import * as guidelines from './building-a-geospatial-app/guidelines';
 
-import GraphLayout from './graph/3-interactive-graph/src/app';
-import InteractiveGraph from './graph/4-final-version/src/app';
+import GraphLayout from './graph/2-graph-layout/src/app';
+import InteractiveGraph from './graph/3-final-version/src/app';
 
 export default {
