+++ src/docs/graph/3-layout-engine.md
@@ -6,12 +6,19 @@
 </ul>
 
 **HOLD UP!!!** If you got here without reading the previous steps,
-it is highly recommended that you do so, or you can just check out the complete code from the last step:
+it is highly recommended that you do so, or you can just check out the complete code from the previous step:
 ```
-cd src/demos/graph/2-graph-layout
+cd src/demos/graph/1-basic-graph
 ```
 
-## D3 Force Layout Engine
+# Architecture
+In this step, we will add a layout engine with the `app.js` for calculating graph layout.
+Here's the overview of the architecture:
+<p class="inline-images">
+  <img src="images/graph-vis/architecture-with-layout-engine.png" alt="extruded" width="600px"/>
+</p>
+
+## 1. D3 Force Layout Engine
 
 First, we need to install [d3-force](https://github.com/d3/d3-force) from npm:
 ```
@@ -63,7 +70,7 @@ Note that, the layout engine should be replaceable with any other implementation
 
 Let's plug in the layout engine we have here with our graph application.
 
-## 1. Start Layout Engine
+## 2. Start Layout Engine
 
 To speed up the rendering, we want the component rerender without 'diffing' the component state of props.
 One trick we can do here is to call 'forceUpdate()' method to rerender the component.
@@ -112,7 +119,7 @@ export default class App extends Component {
 
 See the complete doe of the layout engine at [here]().
 
-## 2. Connect Graph Render with Layout Engine
+## 3. Connect Graph Render with Layout Engine
 
 Once the engine has been launched, we can start to get the node/edge positions from the engine instead of getting the positions from the graph in the component state(`this.state.graph`).
 
@@ -134,7 +141,7 @@ export default class App extends Component {
 }
 ```
 
-## 3. Position Update Trigger
+## 4. Position Update Trigger
 
 <!-- The last part is to trigger `deck.gl` to update when every time the layout is updated. -->
 In the previous step, we connected `getNodePosition` and `getEdgePosition` with the accessors in the layout engine. However, `deck.gl` doesn't recalculate positions unlesss the data prop changes by shallow comparison. To inform deck.gl to re-evaluate `getPosition` outcome, we need to explicitly define `updateTriggers`. 
@@ -194,5 +201,5 @@ export default class App extends Component {
 ## Complete code
 
 You can check the complete code at here:
- - [app.js](https://github.com/uber-common/vis-academy/blob/master/src/demos/graph/3-interactive-graph/src/app.js)
- - [graph-render.js](https://github.com/uber-common/vis-academy/blob/master/src/demos/graph/3-interactive-graph/src/graph-render.js).
+ - [app.js](https://github.com/uber-common/vis-academy/blob/master/src/demos/graph/2-graph-layout/src/app.js)
+ - [graph-render.js](https://github.com/uber-common/vis-academy/blob/master/src/demos/graph/2-graph-layout/src/graph-render.js).
